Add synonyms to GLSLGeneration for 100es and 300es.
We use GLSLGeneration::k330 to mean GLSL 3.30 for desktop GLSL and also
for ES 3.00. k110 lived a double life as ES 1.00. This was confusing, so
several places in the code had comments like "this is the value for GLSL
ES 3" to clarify.

This CL adds new enums "k100es" and "k300es" which use the same values
as "k110" and "k330", but make the code easier to understand at a
glance.

Change-Id: Ib4dfbd8cd7bab45b35d6ad7303ccf87c7eeca3ab
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/506898
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>