Pass squared tolerances into GrPathUtils::generateXPoints()
The GrPathUtils::quadPointCount and cubicPointCount functions take a
tolerance value. The paired generateQuad/CubicPoints functions that will
fill up-to the returned point count from those first questions actually
takes the square of the tolerance value.

This was noticed in https://fiddle.skia.org/c/7e42210ddb6675fe8729c0ba5b07daf5
as the subtle linear artifacts around the rounded corners of the stroke.
GrAAConvexTessellator was not passing in the square value, so the point
generation would early-out sooner than it should have in this example.
Instead of using 16 vertices it used 8. Passing in the square fixes the
linear artifacts when testing locally.

I did an audit of other usages of these functions and SkShadowUtils also
needed to be updated. As part of this, I switched the constants over to
constexpr. The only other user of this function, default path renderer,
did not need to be updated.

Change-Id: I5e0acb8ebc8b4209f7924ca1e1674d1759da0d3e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/507926
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>