add infrastructure for unflattening size for Slugs
First, count the number of sub runs in this Slug.

Second,the Slug is allocated as a large block with an arena at the
end of it. Calculate a hint for allocating that initial large block.

This does not do the calculation for the different sub runs. We are
doing the direct mask case first. Others can be added later.

Change-Id: I80ed206177b50d9bb0742f4d826c3372d12b6fe3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/508639
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>