Give a human-intelligible message if gn generates a warning
If gn generates an otherwise harmless warning, find_headers.py will fail
to parse the gn-generated JSON because gn prints both the warning and
the JSON to stdout. Though find_headers.py prints the output of gn if
JSON parsing fails, a human investigating this failure is likely to
assume that the warning they see is from the build system, not from
find_headers.py, and is not related to the failure.

For example, this warning about an argument which has no effect breaks
the build:

    WARNING at build arg file (use "gn args <out_dir>" to edit):36:36: Build argument has no effect.
    skia_skqp_global_error_tolerance = 8
                                       ^
    [...]

Change-Id: Iafa7252ba161e4def1438f5d9480b64fdaa887d2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/510536
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>