Test removing table lookup from color determination.
In theory this shouldn't make a difference since copying the table
tags should be a super minimal operation (literally reading a few
bytes out of the font header). However, CoreText and CoreGraphics
work in mysterious ways and this may actually be an expensive
operation.

Bug: chromium:1297957
Bug: chromium:1297978
Change-Id: Ia640f4c7469d5d10b3243a5d0e939633a2b6ba7b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/510719
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>