Provide current color to all CoreText color glyphs
While `CPAL` glyphs may explicitly use the current foreground color,
`sbix` glyphs may also need the current foreground color to be correct.
If a font with an `sbix` table contains glyphs with paths but without
an entry in the `sbix` table the glyph may be drawn from the path.
However, it is not possible to know when this might occur with the
current CoreText API, so it is necessary for correctness to provide the
current foreground color for `sbix` fonts as well.

This is actually quite wasteful as it means caching `sbix` glyphs on a
per-current foreground color basis. However, usually the current
foreground color isn't changing for these fonts.

Change-Id: I1c95820d981d463fde673966b31c2c47ab2f3292
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/512147
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>