store GlyphIDs in the PathOpSubmitter
In order to support serializing, have the PathOpSubmitter store
SkPathIDs until the paths are needed for drawing, and then lookup
the paths and store them in the memory that the IDs occupied.
This is a one time operation. Wire the strike down from the
SkGlyphRunPainter down to the PathOpSubmitter.

Change-Id: Id9382fefc252eebafeb9792f380d24ec0939316d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/512937
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Herb Derby <herb@google.com>