In Vulkan readPixels check if mapped buffer returns null.
This is a speculative fix for a crash in Chrome. The memcpy is trying
to use memory at 0x0, so either the dst or src is probably nullptr.
The src is coming from our mapped transfer buffer so this puts a check
in for that. The dst is coming from out in chrome land, but it does
look like they check its valid before calling into Skia.

Bug: chromium:1301769
Change-Id: I10302dc97da6045524ee2f20045816d1f17ccf3d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/514357
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>